refactor: id Int -> Long으로 변경

+++ core/model/src/main/java/com/cmc15th/pluv/core/model/History.kt
@@ -1,7 +1,7 @@
 package com.cmc15th.pluv.core.model
 
 data class History(
-    val id: Int = -1,
+    val id: Long = -1L,
     val transferredSongCount: Int = 0,
     val transferredDate: String = "",
     val title: String = "",
@@ -9,7 +9,7 @@ data class History(
 )
 
 data class HistoryDetail(
-    val id: Int = -1,
+    val id: Long = -1L,
     val totalSongCount: Int = 0,
     val transferredSongCount: Int = 0,
     val title: String = "",
